chore: replace 54 migrations with single initial schema
Drop the full accumulated migration history and replace it with one
autogenerated migration from the current model state. This was the right
time: the DB is being recreated anyway (provider→clinician table rename
means the old DB is incompatible) and the history contained 50+ steps
including many since-dropped columns and renamed tables.

The fresh migration creates all tables with their final correct names
(clinician_certifications, clinician_educations, clinician_licensures,
etc.) and no legacy cruft.
