--- alembic/versions/6dfcb644a327_add_pa_schedule_text.py
@@ -1,36 +1,0 @@
-"""add_pa_schedule_text
-
-Revision ID: 6dfcb644a327
-Revises: 9131c572d8a1
-Create Date: 2026-05-12 07:19:39.867570
-
-"""
-
-from typing import Sequence, Union
-
-import sqlalchemy as sa
-
-from alembic import op
-
-# revision identifiers, used by Alembic.
-revision: str = "6dfcb644a327"
-down_revision: Union[str, None] = "9131c572d8a1"
-branch_labels: Union[str, Sequence[str], None] = None
-depends_on: Union[str, Sequence[str], None] = None
-
-
-def upgrade() -> None:
-    """Upgrade schema."""
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.add_column(
-        "provider_availability_details",
-        sa.Column("schedule_text", sa.Text(), nullable=True),
-    )
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    """Downgrade schema."""
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_column("provider_availability_details", "schedule_text")
-    # ### end Alembic commands ###
