--- alembic/versions/7dbd50c72096_add_provider_profile_licensure_.py
@@ -1,166 +1,0 @@
-"""add provider profile licensure education certification tables
-
-Revision ID: 7dbd50c72096
-Revises: d7e5f0a3c1b9
-Create Date: 2026-05-06 19:11:48.606558
-
-"""
-
-from typing import Sequence, Union
-
-import sqlalchemy as sa
-
-from alembic import op
-
-# revision identifiers, used by Alembic.
-revision: str = "7dbd50c72096"
-down_revision: Union[str, None] = "d7e5f0a3c1b9"
-branch_labels: Union[str, Sequence[str], None] = None
-depends_on: Union[str, Sequence[str], None] = None
-
-
-def upgrade() -> None:
-    """Upgrade schema."""
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.create_table(
-        "provider_profiles",
-        sa.Column("user_id", sa.Uuid(), nullable=False),
-        sa.Column("practice_name", sa.Text(), nullable=False),
-        sa.Column("location_city", sa.Text(), nullable=False),
-        sa.Column("location_state", sa.Text(), nullable=False),
-        sa.Column("location_zip", sa.Text(), nullable=False),
-        sa.Column("in_person_sessions", sa.Text(), nullable=False),
-        sa.Column("virtual_sessions", sa.Text(), nullable=False),
-        sa.Column("id", sa.Uuid(), nullable=False),
-        sa.Column(
-            "created_at",
-            sa.DateTime(timezone=True),
-            server_default=sa.text("(CURRENT_TIMESTAMP)"),
-            nullable=False,
-        ),
-        sa.Column(
-            "updated_at",
-            sa.DateTime(timezone=True),
-            server_default=sa.text("(CURRENT_TIMESTAMP)"),
-            nullable=False,
-        ),
-        sa.Column("deleted_at", sa.DateTime(timezone=True), nullable=True),
-        sa.CheckConstraint(
-            "in_person_sessions IN ('yes', 'no', 'please_contact')",
-            name="ck_provider_profiles_in_person_sessions",
-        ),
-        sa.CheckConstraint(
-            "location_state IN ('AL', 'AK', 'AZ', 'AR', 'CA', 'CO', 'CT', 'DE', 'DC', 'FL', 'GA', 'HI', 'ID', 'IL', 'IN', 'IA', 'KS', 'KY', 'LA', 'ME', 'MD', 'MA', 'MI', 'MN', 'MS', 'MO', 'MT', 'NE', 'NV', 'NH', 'NJ', 'NM', 'NY', 'NC', 'ND', 'OH', 'OK', 'OR', 'PA', 'RI', 'SC', 'SD', 'TN', 'TX', 'UT', 'VT', 'VA', 'WA', 'WV', 'WI', 'WY')",
-            name="ck_provider_profiles_location_state",
-        ),
-        sa.CheckConstraint(
-            "virtual_sessions IN ('yes', 'no', 'please_contact')",
-            name="ck_provider_profiles_virtual_sessions",
-        ),
-        sa.ForeignKeyConstraint(["user_id"], ["users.id"], ondelete="CASCADE"),
-        sa.PrimaryKeyConstraint("id"),
-        sa.UniqueConstraint("user_id", name="uq_provider_profiles_user_id"),
-    )
-    op.create_table(
-        "provider_certifications",
-        sa.Column("profile_id", sa.Uuid(), nullable=False),
-        sa.Column("certification_type", sa.Text(), nullable=False),
-        sa.Column("certifying_body", sa.Text(), nullable=False),
-        sa.Column("expiration_date", sa.Date(), nullable=True),
-        sa.Column("id", sa.Uuid(), nullable=False),
-        sa.Column(
-            "created_at",
-            sa.DateTime(timezone=True),
-            server_default=sa.text("(CURRENT_TIMESTAMP)"),
-            nullable=False,
-        ),
-        sa.Column(
-            "updated_at",
-            sa.DateTime(timezone=True),
-            server_default=sa.text("(CURRENT_TIMESTAMP)"),
-            nullable=False,
-        ),
-        sa.Column("deleted_at", sa.DateTime(timezone=True), nullable=True),
-        sa.CheckConstraint(
-            "certification_type IN ('emdr', 'dbt', 'cbt', 'gottman_1', 'gottman_2', 'gottman_3', 'cpr', 'ccatp', 'other')",
-            name="ck_provider_certifications_certification_type",
-        ),
-        sa.ForeignKeyConstraint(
-            ["profile_id"], ["provider_profiles.id"], ondelete="CASCADE"
-        ),
-        sa.PrimaryKeyConstraint("id"),
-    )
-    op.create_table(
-        "provider_educations",
-        sa.Column("profile_id", sa.Uuid(), nullable=False),
-        sa.Column("education_type", sa.Text(), nullable=False),
-        sa.Column("institution", sa.Text(), nullable=False),
-        sa.Column("month_completed", sa.Text(), nullable=True),
-        sa.Column("id", sa.Uuid(), nullable=False),
-        sa.Column(
-            "created_at",
-            sa.DateTime(timezone=True),
-            server_default=sa.text("(CURRENT_TIMESTAMP)"),
-            nullable=False,
-        ),
-        sa.Column(
-            "updated_at",
-            sa.DateTime(timezone=True),
-            server_default=sa.text("(CURRENT_TIMESTAMP)"),
-            nullable=False,
-        ),
-        sa.Column("deleted_at", sa.DateTime(timezone=True), nullable=True),
-        sa.CheckConstraint(
-            "education_type IN ('ba_bs', 'ma_ms', 'msw', 'phd', 'psyd', 'md', 'do', 'edd', 'other')",
-            name="ck_provider_educations_education_type",
-        ),
-        sa.ForeignKeyConstraint(
-            ["profile_id"], ["provider_profiles.id"], ondelete="CASCADE"
-        ),
-        sa.PrimaryKeyConstraint("id"),
-    )
-    op.create_table(
-        "provider_licensures",
-        sa.Column("profile_id", sa.Uuid(), nullable=False),
-        sa.Column("license_type", sa.Text(), nullable=False),
-        sa.Column("license_number", sa.Text(), nullable=False),
-        sa.Column("issuing_state", sa.Text(), nullable=False),
-        sa.Column("expiration_date", sa.Date(), nullable=True),
-        sa.Column("id", sa.Uuid(), nullable=False),
-        sa.Column(
-            "created_at",
-            sa.DateTime(timezone=True),
-            server_default=sa.text("(CURRENT_TIMESTAMP)"),
-            nullable=False,
-        ),
-        sa.Column(
-            "updated_at",
-            sa.DateTime(timezone=True),
-            server_default=sa.text("(CURRENT_TIMESTAMP)"),
-            nullable=False,
-        ),
-        sa.Column("deleted_at", sa.DateTime(timezone=True), nullable=True),
-        sa.CheckConstraint(
-            "issuing_state IN ('AL', 'AK', 'AZ', 'AR', 'CA', 'CO', 'CT', 'DE', 'DC', 'FL', 'GA', 'HI', 'ID', 'IL', 'IN', 'IA', 'KS', 'KY', 'LA', 'ME', 'MD', 'MA', 'MI', 'MN', 'MS', 'MO', 'MT', 'NE', 'NV', 'NH', 'NJ', 'NM', 'NY', 'NC', 'ND', 'OH', 'OK', 'OR', 'PA', 'RI', 'SC', 'SD', 'TN', 'TX', 'UT', 'VT', 'VA', 'WA', 'WV', 'WI', 'WY')",
-            name="ck_provider_licensures_issuing_state",
-        ),
-        sa.CheckConstraint(
-            "license_type IN ('lcsw', 'lpc', 'lmft', 'lmhc', 'lcpc', 'psyd', 'phd', 'md', 'do', 'np', 'pmhnp', 'other')",
-            name="ck_provider_licensures_license_type",
-        ),
-        sa.ForeignKeyConstraint(
-            ["profile_id"], ["provider_profiles.id"], ondelete="CASCADE"
-        ),
-        sa.PrimaryKeyConstraint("id"),
-    )
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    """Downgrade schema."""
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_table("provider_licensures")
-    op.drop_table("provider_educations")
-    op.drop_table("provider_certifications")
-    op.drop_table("provider_profiles")
-    # ### end Alembic commands ###
