--- alembic/versions/8a6ada1e0883_add_audit_log_table.py
@@ -1,57 +1,0 @@
-"""add audit_log table
-
-Revision ID: 8a6ada1e0883
-Revises: d6d749b73cb7
-Create Date: 2026-04-29 11:17:38.600943
-
-"""
-
-from typing import Sequence, Union
-
-import sqlalchemy as sa
-
-from alembic import op
-
-# revision identifiers, used by Alembic.
-revision: str = "8a6ada1e0883"
-down_revision: Union[str, None] = "d6d749b73cb7"
-branch_labels: Union[str, Sequence[str], None] = None
-depends_on: Union[str, Sequence[str], None] = None
-
-
-def upgrade() -> None:
-    """Upgrade schema."""
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.create_table(
-        "audit_log",
-        sa.Column("actor_id", sa.Uuid(), nullable=True),
-        sa.Column("resource_type", sa.Text(), nullable=False),
-        sa.Column("resource_id", sa.Uuid(), nullable=False),
-        sa.Column("action", sa.Text(), nullable=False),
-        sa.Column("before", sa.JSON(), nullable=True),
-        sa.Column("after", sa.JSON(), nullable=True),
-        sa.Column("id", sa.Uuid(), nullable=False),
-        sa.Column(
-            "created_at",
-            sa.DateTime(timezone=True),
-            server_default=sa.text("(CURRENT_TIMESTAMP)"),
-            nullable=False,
-        ),
-        sa.Column(
-            "updated_at",
-            sa.DateTime(timezone=True),
-            server_default=sa.text("(CURRENT_TIMESTAMP)"),
-            nullable=False,
-        ),
-        sa.Column("deleted_at", sa.DateTime(timezone=True), nullable=True),
-        sa.ForeignKeyConstraint(["actor_id"], ["users.id"], ondelete="SET NULL"),
-        sa.PrimaryKeyConstraint("id"),
-    )
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    """Downgrade schema."""
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_table("audit_log")
-    # ### end Alembic commands ###
