+++ alembic/versions/958dad6b8606_initial_schema.py
@@ -1,0 +1,569 @@
+"""initial schema
+
+Revision ID: 958dad6b8606
+Revises:
+Create Date: 2026-05-27 16:06:23.326834
+
+"""
+
+from typing import Sequence, Union
+
+import sqlalchemy as sa
+
+from alembic import op
+
+# revision identifiers, used by Alembic.
+revision: str = "958dad6b8606"
+down_revision: Union[str, None] = None
+branch_labels: Union[str, Sequence[str], None] = None
+depends_on: Union[str, Sequence[str], None] = None
+
+
+def upgrade() -> None:
+    """Upgrade schema."""
+    # ### commands auto generated by Alembic - please adjust! ###
+    op.create_table(
+        "users",
+        sa.Column("username", sa.Text(), nullable=False),
+        sa.Column("email", sa.String(length=320), nullable=False),
+        sa.Column("hashed_password", sa.String(length=1024), nullable=False),
+        sa.Column("is_active", sa.Boolean(), nullable=False),
+        sa.Column("is_superuser", sa.Boolean(), nullable=False),
+        sa.Column("is_verified", sa.Boolean(), nullable=False),
+        sa.Column("id", sa.Uuid(), nullable=False),
+        sa.Column(
+            "created_at",
+            sa.DateTime(timezone=True),
+            server_default=sa.text("(CURRENT_TIMESTAMP)"),
+            nullable=False,
+        ),
+        sa.Column(
+            "updated_at",
+            sa.DateTime(timezone=True),
+            server_default=sa.text("(CURRENT_TIMESTAMP)"),
+            nullable=False,
+        ),
+        sa.Column("deleted_at", sa.DateTime(timezone=True), nullable=True),
+        sa.PrimaryKeyConstraint("id"),
+        sa.UniqueConstraint("username"),
+    )
+    op.create_index(op.f("ix_users_email"), "users", ["email"], unique=True)
+    op.create_table(
+        "audit_log",
+        sa.Column("actor_id", sa.Uuid(), nullable=True),
+        sa.Column("resource_type", sa.Text(), nullable=False),
+        sa.Column("resource_id", sa.Uuid(), nullable=False),
+        sa.Column("action", sa.Text(), nullable=False),
+        sa.Column("before", sa.JSON(), nullable=True),
+        sa.Column("after", sa.JSON(), nullable=True),
+        sa.Column("id", sa.Uuid(), nullable=False),
+        sa.Column(
+            "created_at",
+            sa.DateTime(timezone=True),
+            server_default=sa.text("(CURRENT_TIMESTAMP)"),
+            nullable=False,
+        ),
+        sa.Column(
+            "updated_at",
+            sa.DateTime(timezone=True),
+            server_default=sa.text("(CURRENT_TIMESTAMP)"),
+            nullable=False,
+        ),
+        sa.Column("deleted_at", sa.DateTime(timezone=True), nullable=True),
+        sa.ForeignKeyConstraint(["actor_id"], ["users.id"], ondelete="SET NULL"),
+        sa.PrimaryKeyConstraint("id"),
+    )
+    op.create_table(
+        "clinicians",
+        sa.Column("owner_id", sa.Uuid(), nullable=True),
+        sa.Column("npi", sa.Text(), nullable=True),
+        sa.Column("first_name", sa.Text(), nullable=True),
+        sa.Column("last_name", sa.Text(), nullable=True),
+        sa.Column("id", sa.Uuid(), nullable=False),
+        sa.Column(
+            "created_at",
+            sa.DateTime(timezone=True),
+            server_default=sa.text("(CURRENT_TIMESTAMP)"),
+            nullable=False,
+        ),
+        sa.Column(
+            "updated_at",
+            sa.DateTime(timezone=True),
+            server_default=sa.text("(CURRENT_TIMESTAMP)"),
+            nullable=False,
+        ),
+        sa.Column("deleted_at", sa.DateTime(timezone=True), nullable=True),
+        sa.CheckConstraint(
+            "npi IS NULL OR (length(npi) = 10 AND npi GLOB '[0-9][0-9][0-9][0-9][0-9][0-9][0-9][0-9][0-9][0-9]')",
+            name="ck_clinicians_npi_format",
+        ),
+        sa.ForeignKeyConstraint(["owner_id"], ["users.id"], ondelete="CASCADE"),
+        sa.PrimaryKeyConstraint("id"),
+    )
+    op.create_table(
+        "organizations",
+        sa.Column("name", sa.Text(), nullable=False),
+        sa.Column("type", sa.Text(), nullable=False),
+        sa.Column("owner_id", sa.Uuid(), nullable=False),
+        sa.Column("parent_org_id", sa.Uuid(), nullable=True),
+        sa.Column("root_org_id", sa.Uuid(), nullable=False),
+        sa.Column("id", sa.Uuid(), nullable=False),
+        sa.Column(
+            "created_at",
+            sa.DateTime(timezone=True),
+            server_default=sa.text("(CURRENT_TIMESTAMP)"),
+            nullable=False,
+        ),
+        sa.Column(
+            "updated_at",
+            sa.DateTime(timezone=True),
+            server_default=sa.text("(CURRENT_TIMESTAMP)"),
+            nullable=False,
+        ),
+        sa.Column("deleted_at", sa.DateTime(timezone=True), nullable=True),
+        sa.CheckConstraint(
+            "type IN ('solo_practice', 'group_practice', 'clinic', 'health_system', 'other')",
+            name="ck_organizations_type",
+        ),
+        sa.ForeignKeyConstraint(["owner_id"], ["users.id"], ondelete="CASCADE"),
+        sa.ForeignKeyConstraint(
+            ["parent_org_id"], ["organizations.id"], ondelete="CASCADE"
+        ),
+        sa.ForeignKeyConstraint(
+            ["root_org_id"], ["organizations.id"], ondelete="CASCADE"
+        ),
+        sa.PrimaryKeyConstraint("id"),
+    )
+    op.create_table(
+        "posts",
+        sa.Column("kind", sa.Text(), nullable=False),
+        sa.Column("owner_id", sa.Uuid(), nullable=False),
+        sa.Column("id", sa.Uuid(), nullable=False),
+        sa.Column(
+            "created_at",
+            sa.DateTime(timezone=True),
+            server_default=sa.text("(CURRENT_TIMESTAMP)"),
+            nullable=False,
+        ),
+        sa.Column(
+            "updated_at",
+            sa.DateTime(timezone=True),
+            server_default=sa.text("(CURRENT_TIMESTAMP)"),
+            nullable=False,
+        ),
+        sa.Column("deleted_at", sa.DateTime(timezone=True), nullable=True),
+        sa.CheckConstraint(
+            "kind IN ('referral', 'clinician_opening', 'program_intake')",
+            name="ck_posts_kind",
+        ),
+        sa.ForeignKeyConstraint(["owner_id"], ["users.id"], ondelete="CASCADE"),
+        sa.PrimaryKeyConstraint("id"),
+    )
+    op.create_table(
+        "affiliations",
+        sa.Column("clinician_id", sa.Uuid(), nullable=False),
+        sa.Column("org_id", sa.Uuid(), nullable=False),
+        sa.Column("in_person_sessions", sa.Text(), nullable=False),
+        sa.Column("virtual_sessions", sa.Text(), nullable=False),
+        sa.Column(
+            "accepts_out_of_network",
+            sa.Boolean(),
+            server_default=sa.text("1"),
+            nullable=False,
+        ),
+        sa.Column(
+            "in_network_carriers",
+            sa.JSON(),
+            server_default=sa.text("'[]'"),
+            nullable=False,
+        ),
+        sa.Column(
+            "sliding_scale", sa.Boolean(), server_default=sa.text("0"), nullable=False
+        ),
+        sa.Column("cost", sa.Text(), nullable=True),
+        sa.Column("location_city", sa.Text(), nullable=False),
+        sa.Column("location_state", sa.Text(), nullable=False),
+        sa.Column("location_zip", sa.Text(), nullable=False),
+        sa.Column("id", sa.Uuid(), nullable=False),
+        sa.Column(
+            "created_at",
+            sa.DateTime(timezone=True),
+            server_default=sa.text("(CURRENT_TIMESTAMP)"),
+            nullable=False,
+        ),
+        sa.Column(
+            "updated_at",
+            sa.DateTime(timezone=True),
+            server_default=sa.text("(CURRENT_TIMESTAMP)"),
+            nullable=False,
+        ),
+        sa.Column("deleted_at", sa.DateTime(timezone=True), nullable=True),
+        sa.CheckConstraint(
+            "in_person_sessions IN ('yes', 'no', 'please_contact')",
+            name="ck_affiliations_in_person_sessions",
+        ),
+        sa.CheckConstraint(
+            "location_state IN ('AL', 'AK', 'AZ', 'AR', 'CA', 'CO', 'CT', 'DE', 'DC', 'FL', 'GA', 'HI', 'ID', 'IL', 'IN', 'IA', 'KS', 'KY', 'LA', 'ME', 'MD', 'MA', 'MI', 'MN', 'MS', 'MO', 'MT', 'NE', 'NV', 'NH', 'NJ', 'NM', 'NY', 'NC', 'ND', 'OH', 'OK', 'OR', 'PA', 'RI', 'SC', 'SD', 'TN', 'TX', 'UT', 'VT', 'VA', 'WA', 'WV', 'WI', 'WY')",
+            name="ck_affiliations_location_state",
+        ),
+        sa.CheckConstraint(
+            "virtual_sessions IN ('yes', 'no', 'please_contact')",
+            name="ck_affiliations_virtual_sessions",
+        ),
+        sa.ForeignKeyConstraint(
+            ["clinician_id"], ["clinicians.id"], ondelete="RESTRICT"
+        ),
+        sa.ForeignKeyConstraint(["org_id"], ["organizations.id"], ondelete="RESTRICT"),
+        sa.PrimaryKeyConstraint("id"),
+    )
+    op.create_table(
+        "clinician_certifications",
+        sa.Column("clinician_id", sa.Uuid(), nullable=False),
+        sa.Column("certification_type", sa.Text(), nullable=False),
+        sa.Column("certifying_body", sa.Text(), nullable=False),
+        sa.Column("expiration_date", sa.Date(), nullable=True),
+        sa.Column("id", sa.Uuid(), nullable=False),
+        sa.Column(
+            "created_at",
+            sa.DateTime(timezone=True),
+            server_default=sa.text("(CURRENT_TIMESTAMP)"),
+            nullable=False,
+        ),
+        sa.Column(
+            "updated_at",
+            sa.DateTime(timezone=True),
+            server_default=sa.text("(CURRENT_TIMESTAMP)"),
+            nullable=False,
+        ),
+        sa.Column("deleted_at", sa.DateTime(timezone=True), nullable=True),
+        sa.CheckConstraint(
+            "certification_type IN ('emdr', 'dbt', 'cbt', 'gottman_1', 'gottman_2', 'gottman_3', 'cpr', 'ccatp', 'other')",
+            name="ck_clinician_certifications_certification_type",
+        ),
+        sa.ForeignKeyConstraint(
+            ["clinician_id"], ["clinicians.id"], ondelete="CASCADE"
+        ),
+        sa.PrimaryKeyConstraint("id"),
+    )
+    op.create_table(
+        "clinician_educations",
+        sa.Column("clinician_id", sa.Uuid(), nullable=False),
+        sa.Column("education_type", sa.Text(), nullable=False),
+        sa.Column("institution", sa.Text(), nullable=False),
+        sa.Column("month_completed", sa.Text(), nullable=True),
+        sa.Column("id", sa.Uuid(), nullable=False),
+        sa.Column(
+            "created_at",
+            sa.DateTime(timezone=True),
+            server_default=sa.text("(CURRENT_TIMESTAMP)"),
+            nullable=False,
+        ),
+        sa.Column(
+            "updated_at",
+            sa.DateTime(timezone=True),
+            server_default=sa.text("(CURRENT_TIMESTAMP)"),
+            nullable=False,
+        ),
+        sa.Column("deleted_at", sa.DateTime(timezone=True), nullable=True),
+        sa.CheckConstraint(
+            "education_type IN ('ba_bs', 'ma_ms', 'msw', 'phd', 'psyd', 'md', 'do', 'edd', 'other')",
+            name="ck_clinician_educations_education_type",
+        ),
+        sa.ForeignKeyConstraint(
+            ["clinician_id"], ["clinicians.id"], ondelete="CASCADE"
+        ),
+        sa.PrimaryKeyConstraint("id"),
+    )
+    op.create_table(
+        "clinician_licensures",
+        sa.Column("clinician_id", sa.Uuid(), nullable=False),
+        sa.Column("license_type", sa.Text(), nullable=False),
+        sa.Column("license_number", sa.Text(), nullable=False),
+        sa.Column("issuing_state", sa.Text(), nullable=False),
+        sa.Column("expiration_date", sa.Date(), nullable=True),
+        sa.Column("id", sa.Uuid(), nullable=False),
+        sa.Column(
+            "created_at",
+            sa.DateTime(timezone=True),
+            server_default=sa.text("(CURRENT_TIMESTAMP)"),
+            nullable=False,
+        ),
+        sa.Column(
+            "updated_at",
+            sa.DateTime(timezone=True),
+            server_default=sa.text("(CURRENT_TIMESTAMP)"),
+            nullable=False,
+        ),
+        sa.Column("deleted_at", sa.DateTime(timezone=True), nullable=True),
+        sa.CheckConstraint(
+            "issuing_state IN ('AL', 'AK', 'AZ', 'AR', 'CA', 'CO', 'CT', 'DE', 'DC', 'FL', 'GA', 'HI', 'ID', 'IL', 'IN', 'IA', 'KS', 'KY', 'LA', 'ME', 'MD', 'MA', 'MI', 'MN', 'MS', 'MO', 'MT', 'NE', 'NV', 'NH', 'NJ', 'NM', 'NY', 'NC', 'ND', 'OH', 'OK', 'OR', 'PA', 'RI', 'SC', 'SD', 'TN', 'TX', 'UT', 'VT', 'VA', 'WA', 'WV', 'WI', 'WY')",
+            name="ck_clinician_licensures_issuing_state",
+        ),
+        sa.CheckConstraint(
+            "license_type IN ('lcsw', 'lpc', 'lmft', 'lmhc', 'lcpc', 'psyd', 'phd', 'md', 'do', 'np', 'pmhnp', 'other')",
+            name="ck_clinician_licensures_license_type",
+        ),
+        sa.ForeignKeyConstraint(
+            ["clinician_id"], ["clinicians.id"], ondelete="CASCADE"
+        ),
+        sa.PrimaryKeyConstraint("id"),
+    )
+    op.create_table(
+        "opening_details",
+        sa.Column("post_id", sa.Uuid(), nullable=False),
+        sa.Column("clinician_id", sa.Uuid(), nullable=False),
+        sa.Column(
+            "desired_times", sa.JSON(), server_default=sa.text("'[]'"), nullable=False
+        ),
+        sa.Column("schedule_text", sa.Text(), nullable=True),
+        sa.Column(
+            "services", sa.JSON(), server_default=sa.text("'[]'"), nullable=False
+        ),
+        sa.Column(
+            "settings", sa.JSON(), server_default=sa.text("'[]'"), nullable=False
+        ),
+        sa.Column("treatment_modality", sa.Text(), nullable=True),
+        sa.Column(
+            "age_groups", sa.JSON(), server_default=sa.text("'[]'"), nullable=False
+        ),
+        sa.Column(
+            "languages", sa.JSON(), server_default=sa.text("'[\"en\"]'"), nullable=False
+        ),
+        sa.Column("genders", sa.JSON(), server_default=sa.text("'[]'"), nullable=False),
+        sa.Column("subject", sa.Text(), nullable=True),
+        sa.Column("description", sa.Text(), nullable=True),
+        sa.Column("referral_instructions", sa.Text(), nullable=True),
+        sa.Column("website", sa.Text(), nullable=True),
+        sa.ForeignKeyConstraint(
+            ["clinician_id"], ["clinicians.id"], ondelete="CASCADE"
+        ),
+        sa.ForeignKeyConstraint(["post_id"], ["posts.id"], ondelete="CASCADE"),
+        sa.PrimaryKeyConstraint("post_id"),
+    )
+    op.create_table(
+        "programs",
+        sa.Column("owner_id", sa.Uuid(), nullable=False),
+        sa.Column("org_id", sa.Uuid(), nullable=False),
+        sa.Column("name", sa.Text(), nullable=False),
+        sa.Column("description", sa.Text(), nullable=True),
+        sa.Column("state_preference", sa.Text(), nullable=True),
+        sa.Column("start_date", sa.Date(), nullable=True),
+        sa.Column("end_date", sa.Date(), nullable=True),
+        sa.Column(
+            "accepting_referrals",
+            sa.Boolean(),
+            server_default=sa.text("1"),
+            nullable=False,
+        ),
+        sa.Column("id", sa.Uuid(), nullable=False),
+        sa.Column(
+            "created_at",
+            sa.DateTime(timezone=True),
+            server_default=sa.text("(CURRENT_TIMESTAMP)"),
+            nullable=False,
+        ),
+        sa.Column(
+            "updated_at",
+            sa.DateTime(timezone=True),
+            server_default=sa.text("(CURRENT_TIMESTAMP)"),
+            nullable=False,
+        ),
+        sa.Column("deleted_at", sa.DateTime(timezone=True), nullable=True),
+        sa.CheckConstraint(
+            "state_preference IN ('AL', 'AK', 'AZ', 'AR', 'CA', 'CO', 'CT', 'DE', 'DC', 'FL', 'GA', 'HI', 'ID', 'IL', 'IN', 'IA', 'KS', 'KY', 'LA', 'ME', 'MD', 'MA', 'MI', 'MN', 'MS', 'MO', 'MT', 'NE', 'NV', 'NH', 'NJ', 'NM', 'NY', 'NC', 'ND', 'OH', 'OK', 'OR', 'PA', 'RI', 'SC', 'SD', 'TN', 'TX', 'UT', 'VT', 'VA', 'WA', 'WV', 'WI', 'WY')",
+            name="ck_programs_state_preference",
+        ),
+        sa.ForeignKeyConstraint(["org_id"], ["organizations.id"], ondelete="RESTRICT"),
+        sa.ForeignKeyConstraint(["owner_id"], ["users.id"], ondelete="CASCADE"),
+        sa.PrimaryKeyConstraint("id"),
+    )
+    op.create_table(
+        "referral_details",
+        sa.Column("post_id", sa.Uuid(), nullable=False),
+        sa.Column("location_in_person", sa.Text(), nullable=False),
+        sa.Column("location_virtual", sa.Text(), nullable=False),
+        sa.Column(
+            "desired_times", sa.JSON(), server_default=sa.text("'[]'"), nullable=False
+        ),
+        sa.Column(
+            "age_groups", sa.JSON(), server_default=sa.text("'[]'"), nullable=False
+        ),
+        sa.Column(
+            "languages", sa.JSON(), server_default=sa.text("'[\"en\"]'"), nullable=False
+        ),
+        sa.Column(
+            "gender",
+            sa.Text(),
+            server_default=sa.text("'prefer_not_to_say'"),
+            nullable=False,
+        ),
+        sa.Column("subject", sa.Text(), nullable=True),
+        sa.Column("description", sa.Text(), nullable=False),
+        sa.Column(
+            "services", sa.JSON(), server_default=sa.text("'[]'"), nullable=False
+        ),
+        sa.Column("treatment_modality", sa.Text(), nullable=True),
+        sa.Column(
+            "network_preference",
+            sa.Text(),
+            server_default=sa.text("'no_preference'"),
+            nullable=False,
+        ),
+        sa.Column("insurance_carrier", sa.Text(), nullable=True),
+        sa.Column("location_city", sa.Text(), nullable=False),
+        sa.Column("location_state", sa.Text(), nullable=False),
+        sa.Column("location_zip", sa.Text(), nullable=False),
+        sa.CheckConstraint(
+            "gender IN ('female', 'male', 'non_binary', 'trans_female', 'trans_male', 'gender_diverse', 'prefer_not_to_say')",
+            name="ck_referral_details_gender",
+        ),
+        sa.CheckConstraint(
+            "insurance_carrier IN ('aetna', 'anthem_bcbs', 'cigna', 'kaiser', 'magellan', 'medicare', 'medicaid', 'optum', 'tricare', 'united_healthcare', 'other')",
+            name="ck_referral_details_insurance_carrier",
+        ),
+        sa.CheckConstraint(
+            "location_in_person IN ('yes', 'no', 'please_contact')",
+            name="ck_referral_details_location_in_person",
+        ),
+        sa.CheckConstraint(
+            "location_state IN ('AL', 'AK', 'AZ', 'AR', 'CA', 'CO', 'CT', 'DE', 'DC', 'FL', 'GA', 'HI', 'ID', 'IL', 'IN', 'IA', 'KS', 'KY', 'LA', 'ME', 'MD', 'MA', 'MI', 'MN', 'MS', 'MO', 'MT', 'NE', 'NV', 'NH', 'NJ', 'NM', 'NY', 'NC', 'ND', 'OH', 'OK', 'OR', 'PA', 'RI', 'SC', 'SD', 'TN', 'TX', 'UT', 'VT', 'VA', 'WA', 'WV', 'WI', 'WY')",
+            name="ck_referral_details_location_state",
+        ),
+        sa.CheckConstraint(
+            "location_virtual IN ('yes', 'no', 'please_contact')",
+            name="ck_referral_details_location_virtual",
+        ),
+        sa.CheckConstraint(
+            "network_preference IN ('in_network_required', 'in_network_preferred', 'no_preference')",
+            name="ck_referral_details_network_preference",
+        ),
+        sa.ForeignKeyConstraint(["post_id"], ["posts.id"], ondelete="CASCADE"),
+        sa.PrimaryKeyConstraint("post_id"),
+    )
+    op.create_table(
+        "user_favorites",
+        sa.Column("user_id", sa.Uuid(), nullable=False),
+        sa.Column("clinician_id", sa.Uuid(), nullable=False),
+        sa.Column("id", sa.Uuid(), nullable=False),
+        sa.Column(
+            "created_at",
+            sa.DateTime(timezone=True),
+            server_default=sa.text("(CURRENT_TIMESTAMP)"),
+            nullable=False,
+        ),
+        sa.Column(
+            "updated_at",
+            sa.DateTime(timezone=True),
+            server_default=sa.text("(CURRENT_TIMESTAMP)"),
+            nullable=False,
+        ),
+        sa.Column("deleted_at", sa.DateTime(timezone=True), nullable=True),
+        sa.ForeignKeyConstraint(
+            ["clinician_id"], ["clinicians.id"], ondelete="CASCADE"
+        ),
+        sa.ForeignKeyConstraint(["user_id"], ["users.id"], ondelete="CASCADE"),
+        sa.PrimaryKeyConstraint("id"),
+        sa.UniqueConstraint(
+            "user_id", "clinician_id", name="uq_user_favorites_user_clinician"
+        ),
+    )
+    op.create_index(
+        "ix_user_favorites_user_id_created_at",
+        "user_favorites",
+        ["user_id", "created_at"],
+        unique=False,
+    )
+    op.create_table(
+        "verifications",
+        sa.Column("clinician_id", sa.Uuid(), nullable=False),
+        sa.Column("status", sa.Text(), nullable=False),
+        sa.Column("flags", sa.JSON(), server_default=sa.text("'[]'"), nullable=False),
+        sa.Column("nppes_result", sa.JSON(), nullable=True),
+        sa.Column(
+            "oig_match", sa.Boolean(), server_default=sa.text("0"), nullable=False
+        ),
+        sa.Column("name_match_score", sa.Float(), nullable=True),
+        sa.Column("id", sa.Uuid(), nullable=False),
+        sa.Column(
+            "created_at",
+            sa.DateTime(timezone=True),
+            server_default=sa.text("(CURRENT_TIMESTAMP)"),
+            nullable=False,
+        ),
+        sa.Column(
+            "updated_at",
+            sa.DateTime(timezone=True),
+            server_default=sa.text("(CURRENT_TIMESTAMP)"),
+            nullable=False,
+        ),
+        sa.Column("deleted_at", sa.DateTime(timezone=True), nullable=True),
+        sa.CheckConstraint(
+            "status IN ('verified', 'needs_review', 'failed')",
+            name="ck_verifications_status",
+        ),
+        sa.ForeignKeyConstraint(
+            ["clinician_id"], ["clinicians.id"], ondelete="CASCADE"
+        ),
+        sa.PrimaryKeyConstraint("id"),
+    )
+    op.create_index(
+        op.f("ix_verifications_clinician_id"),
+        "verifications",
+        ["clinician_id"],
+        unique=False,
+    )
+    op.create_table(
+        "intake_details",
+        sa.Column("post_id", sa.Uuid(), nullable=False),
+        sa.Column("program_id", sa.Uuid(), nullable=False),
+        sa.Column(
+            "desired_times", sa.JSON(), server_default=sa.text("'[]'"), nullable=False
+        ),
+        sa.Column("schedule_text", sa.Text(), nullable=True),
+        sa.Column(
+            "services", sa.JSON(), server_default=sa.text("'[]'"), nullable=False
+        ),
+        sa.Column(
+            "settings", sa.JSON(), server_default=sa.text("'[]'"), nullable=False
+        ),
+        sa.Column("treatment_modality", sa.Text(), nullable=True),
+        sa.Column(
+            "age_groups", sa.JSON(), server_default=sa.text("'[]'"), nullable=False
+        ),
+        sa.Column(
+            "languages", sa.JSON(), server_default=sa.text("'[\"en\"]'"), nullable=False
+        ),
+        sa.Column("genders", sa.JSON(), server_default=sa.text("'[]'"), nullable=False),
+        sa.Column("subject", sa.Text(), nullable=True),
+        sa.Column("description", sa.Text(), nullable=True),
+        sa.Column("referral_instructions", sa.Text(), nullable=True),
+        sa.Column("website", sa.Text(), nullable=True),
+        sa.ForeignKeyConstraint(["post_id"], ["posts.id"], ondelete="CASCADE"),
+        sa.ForeignKeyConstraint(["program_id"], ["programs.id"], ondelete="CASCADE"),
+        sa.PrimaryKeyConstraint("post_id"),
+    )
+    # ### end Alembic commands ###
+
+
+def downgrade() -> None:
+    """Downgrade schema."""
+    # ### commands auto generated by Alembic - please adjust! ###
+    op.drop_table("intake_details")
+    op.drop_index(op.f("ix_verifications_clinician_id"), table_name="verifications")
+    op.drop_table("verifications")
+    op.drop_index("ix_user_favorites_user_id_created_at", table_name="user_favorites")
+    op.drop_table("user_favorites")
+    op.drop_table("referral_details")
+    op.drop_table("programs")
+    op.drop_table("opening_details")
+    op.drop_table("clinician_licensures")
+    op.drop_table("clinician_educations")
+    op.drop_table("clinician_certifications")
+    op.drop_table("affiliations")
+    op.drop_table("posts")
+    op.drop_table("organizations")
+    op.drop_table("clinicians")
+    op.drop_table("audit_log")
+    op.drop_index(op.f("ix_users_email"), table_name="users")
+    op.drop_table("users")
+    # ### end Alembic commands ###
