--- alembic/versions/a50977e5627b_add_pa_free_text_fields.py
@@ -1,45 +1,0 @@
-"""add_pa_free_text_fields
-
-Revision ID: a50977e5627b
-Revises: ff47c5267ea2
-Create Date: 2026-05-11 20:53:51.173981
-
-"""
-
-from typing import Sequence, Union
-
-import sqlalchemy as sa
-
-from alembic import op
-
-# revision identifiers, used by Alembic.
-revision: str = "a50977e5627b"
-down_revision: Union[str, None] = "ff47c5267ea2"
-branch_labels: Union[str, Sequence[str], None] = None
-depends_on: Union[str, Sequence[str], None] = None
-
-
-def upgrade() -> None:
-    """Upgrade schema."""
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.add_column(
-        "provider_availability_details",
-        sa.Column("description", sa.Text(), nullable=True),
-    )
-    op.add_column(
-        "provider_availability_details",
-        sa.Column("referral_instructions", sa.Text(), nullable=True),
-    )
-    op.add_column(
-        "provider_availability_details", sa.Column("website", sa.Text(), nullable=True)
-    )
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    """Downgrade schema."""
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_column("provider_availability_details", "website")
-    op.drop_column("provider_availability_details", "referral_instructions")
-    op.drop_column("provider_availability_details", "description")
-    # ### end Alembic commands ###
