--- alembic/versions/b2dbe3136387_drop_chat_tables_and_columns.py
@@ -1,142 +1,0 @@
-"""drop_chat_tables_and_columns
-
-Revision ID: b2dbe3136387
-Revises: 615021cc5b23
-Create Date: 2026-04-23 18:03:19.415660
-
-"""
-
-from typing import Sequence, Union
-
-import sqlalchemy as sa
-
-from alembic import op
-
-# revision identifiers, used by Alembic.
-revision: str = "b2dbe3136387"
-down_revision: Union[str, None] = "615021cc5b23"
-branch_labels: Union[str, Sequence[str], None] = None
-depends_on: Union[str, Sequence[str], None] = None
-
-
-def upgrade() -> None:
-    """Upgrade schema."""
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_table("conversations")
-    op.drop_table("messages")
-    op.drop_table("participants")
-    op.drop_column("users", "last_active_at")
-    op.drop_column("users", "is_online")
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    """Downgrade schema."""
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.add_column(
-        "users",
-        sa.Column(
-            "is_online", sa.BOOLEAN(), server_default=sa.text("0"), nullable=False
-        ),
-    )
-    op.add_column("users", sa.Column("last_active_at", sa.DATETIME(), nullable=True))
-    op.create_table(
-        "participants",
-        sa.Column("user_id", sa.CHAR(length=32), nullable=False),
-        sa.Column("conversation_id", sa.CHAR(length=32), nullable=False),
-        sa.Column("status", sa.VARCHAR(length=8), nullable=False),
-        sa.Column("invited_by_user_id", sa.CHAR(length=32), nullable=True),
-        sa.Column("initial_message_id", sa.CHAR(length=32), nullable=True),
-        sa.Column("joined_at", sa.DATETIME(), nullable=True),
-        sa.Column("id", sa.CHAR(length=32), nullable=False),
-        sa.Column(
-            "created_at",
-            sa.DATETIME(),
-            server_default=sa.text("(CURRENT_TIMESTAMP)"),
-            nullable=False,
-        ),
-        sa.Column(
-            "updated_at",
-            sa.DATETIME(),
-            server_default=sa.text("(CURRENT_TIMESTAMP)"),
-            nullable=False,
-        ),
-        sa.Column("deleted_at", sa.DATETIME(), nullable=True),
-        sa.ForeignKeyConstraint(
-            ["conversation_id"],
-            ["conversations.id"],
-        ),
-        sa.ForeignKeyConstraint(
-            ["initial_message_id"],
-            ["messages.id"],
-        ),
-        sa.ForeignKeyConstraint(
-            ["invited_by_user_id"],
-            ["users.id"],
-        ),
-        sa.ForeignKeyConstraint(
-            ["user_id"],
-            ["users.id"],
-        ),
-        sa.PrimaryKeyConstraint("id"),
-        sa.UniqueConstraint(
-            "user_id", "conversation_id", name="uq_participant_user_conversation"
-        ),
-    )
-    op.create_table(
-        "messages",
-        sa.Column("content", sa.TEXT(), nullable=False),
-        sa.Column("conversation_id", sa.CHAR(length=32), nullable=False),
-        sa.Column("created_by_user_id", sa.CHAR(length=32), nullable=False),
-        sa.Column("id", sa.CHAR(length=32), nullable=False),
-        sa.Column(
-            "created_at",
-            sa.DATETIME(),
-            server_default=sa.text("(CURRENT_TIMESTAMP)"),
-            nullable=False,
-        ),
-        sa.Column(
-            "updated_at",
-            sa.DATETIME(),
-            server_default=sa.text("(CURRENT_TIMESTAMP)"),
-            nullable=False,
-        ),
-        sa.Column("deleted_at", sa.DATETIME(), nullable=True),
-        sa.ForeignKeyConstraint(
-            ["conversation_id"],
-            ["conversations.id"],
-        ),
-        sa.ForeignKeyConstraint(
-            ["created_by_user_id"],
-            ["users.id"],
-        ),
-        sa.PrimaryKeyConstraint("id"),
-    )
-    op.create_table(
-        "conversations",
-        sa.Column("name", sa.TEXT(), nullable=True),
-        sa.Column("slug", sa.TEXT(), nullable=False),
-        sa.Column("created_by_user_id", sa.CHAR(length=32), nullable=False),
-        sa.Column("last_activity_at", sa.DATETIME(), nullable=True),
-        sa.Column("id", sa.CHAR(length=32), nullable=False),
-        sa.Column(
-            "created_at",
-            sa.DATETIME(),
-            server_default=sa.text("(CURRENT_TIMESTAMP)"),
-            nullable=False,
-        ),
-        sa.Column(
-            "updated_at",
-            sa.DATETIME(),
-            server_default=sa.text("(CURRENT_TIMESTAMP)"),
-            nullable=False,
-        ),
-        sa.Column("deleted_at", sa.DATETIME(), nullable=True),
-        sa.ForeignKeyConstraint(
-            ["created_by_user_id"],
-            ["users.id"],
-        ),
-        sa.PrimaryKeyConstraint("id"),
-        sa.UniqueConstraint("slug"),
-    )
-    # ### end Alembic commands ###
