--- alembic/versions/f89c9e2a7748_add_verifications_table.py
@@ -1,70 +1,0 @@
-"""add verifications table
-
-Revision ID: f89c9e2a7748
-Revises: 779078fd2e23
-Create Date: 2026-05-17 21:31:15.367227
-
-"""
-
-from typing import Sequence, Union
-
-import sqlalchemy as sa
-
-from alembic import op
-
-# revision identifiers, used by Alembic.
-revision: str = "f89c9e2a7748"
-down_revision: Union[str, None] = "2cc8e54d7a19"
-branch_labels: Union[str, Sequence[str], None] = None
-depends_on: Union[str, Sequence[str], None] = None
-
-
-def upgrade() -> None:
-    """Upgrade schema."""
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.create_table(
-        "verifications",
-        sa.Column("provider_id", sa.Uuid(), nullable=False),
-        sa.Column("status", sa.Text(), nullable=False),
-        sa.Column("flags", sa.JSON(), server_default=sa.text("'[]'"), nullable=False),
-        sa.Column("nppes_result", sa.JSON(), nullable=True),
-        sa.Column(
-            "oig_match", sa.Boolean(), server_default=sa.text("0"), nullable=False
-        ),
-        sa.Column("name_match_score", sa.Float(), nullable=True),
-        sa.Column("id", sa.Uuid(), nullable=False),
-        sa.Column(
-            "created_at",
-            sa.DateTime(timezone=True),
-            server_default=sa.text("(CURRENT_TIMESTAMP)"),
-            nullable=False,
-        ),
-        sa.Column(
-            "updated_at",
-            sa.DateTime(timezone=True),
-            server_default=sa.text("(CURRENT_TIMESTAMP)"),
-            nullable=False,
-        ),
-        sa.Column("deleted_at", sa.DateTime(timezone=True), nullable=True),
-        sa.CheckConstraint(
-            "status IN ('verified', 'needs_review', 'failed')",
-            name="ck_verifications_status",
-        ),
-        sa.ForeignKeyConstraint(["provider_id"], ["providers.id"], ondelete="CASCADE"),
-        sa.PrimaryKeyConstraint("id"),
-    )
-    op.create_index(
-        op.f("ix_verifications_provider_id"),
-        "verifications",
-        ["provider_id"],
-        unique=False,
-    )
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    """Downgrade schema."""
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_index(op.f("ix_verifications_provider_id"), table_name="verifications")
-    op.drop_table("verifications")
-    # ### end Alembic commands ###
