--- alembic/versions/ff47c5267ea2_add_user_favorites_table.py
@@ -1,64 +1,0 @@
-"""Add user_favorites table
-
-Revision ID: ff47c5267ea2
-Revises: e24d955563a5
-Create Date: 2026-05-10 14:03:58.514053
-
-"""
-
-from typing import Sequence, Union
-
-import sqlalchemy as sa
-
-from alembic import op
-
-# revision identifiers, used by Alembic.
-revision: str = "ff47c5267ea2"
-down_revision: Union[str, None] = "e24d955563a5"
-branch_labels: Union[str, Sequence[str], None] = None
-depends_on: Union[str, Sequence[str], None] = None
-
-
-def upgrade() -> None:
-    """Upgrade schema."""
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.create_table(
-        "user_favorites",
-        sa.Column("user_id", sa.Uuid(), nullable=False),
-        sa.Column("provider_id", sa.Uuid(), nullable=False),
-        sa.Column("id", sa.Uuid(), nullable=False),
-        sa.Column(
-            "created_at",
-            sa.DateTime(timezone=True),
-            server_default=sa.text("(CURRENT_TIMESTAMP)"),
-            nullable=False,
-        ),
-        sa.Column(
-            "updated_at",
-            sa.DateTime(timezone=True),
-            server_default=sa.text("(CURRENT_TIMESTAMP)"),
-            nullable=False,
-        ),
-        sa.Column("deleted_at", sa.DateTime(timezone=True), nullable=True),
-        sa.ForeignKeyConstraint(["provider_id"], ["providers.id"], ondelete="CASCADE"),
-        sa.ForeignKeyConstraint(["user_id"], ["users.id"], ondelete="CASCADE"),
-        sa.PrimaryKeyConstraint("id"),
-        sa.UniqueConstraint(
-            "user_id", "provider_id", name="uq_user_favorites_user_provider"
-        ),
-    )
-    op.create_index(
-        "ix_user_favorites_user_id_created_at",
-        "user_favorites",
-        ["user_id", "created_at"],
-        unique=False,
-    )
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    """Downgrade schema."""
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_index("ix_user_favorites_user_id_created_at", table_name="user_favorites")
-    op.drop_table("user_favorites")
-    # ### end Alembic commands ###
